backend/requirements.txt
asgiref==3.11.0
camelot-py==0.11.0
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
et_xmlfile==2.0.0
fire==0.7.1
fonttools==4.61.1
ghostscript==0.8.1
gunicorn==23.0.0
lxml==6.0.2
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2docx==0.5.6
pdfminer.six==20251227
pdftopng==0.2.4
Pillow==10.1.0
pycparser==2.23
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pypdf==6.5.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
tabulate==0.9.0
termcolor==3.2.0
typing_extensions==4.15.0
tzdata==2025.3
